correcoes de bugs

+++ src/firmware/odometria/components/encoder/encoder.c
@@ -116,7 +116,7 @@ float encoder_get_teta(encoder_t *encoder){
 }
 
 //velocidade angular media [rad/s]
-float encoder_get_w(encoder_t *encoder, int64_t dt){
+float encoder_get_w(encoder_t *encoder, float dt){
     
     float dteta = encoder_get_teta(encoder);
 
@@ -125,9 +125,8 @@ float encoder_get_w(encoder_t *encoder, int64_t dt){
     return w;
 }
 
-
 //velocidade linear media [m/s]
-float encoder_get_v(encoder_t *encoder, int64_t dt, float raio){
+float encoder_get_v(encoder_t *encoder, float dt, float raio){
     float w = encoder_get_w(encoder, dt);
     float v = w * raio;
